OLH-2970: Deploy the VPC template to all accounts
Note I've set `AllowRules` and `AllowedDomains` to allow outbound
traffic from these VPCs to our domains. This is because we'll need to be
able to reach our clients' JWKS endpoints so we can verfy their signing
keys.

`*.account.gov.uk` is included in the default rules for the domain and
network firewalls, but if both these parameters are empty then the
template won't configure the VPC to allow outbound traffic.