OLH-3249: Language switcher tweaks
Add language detector middleware which automatically sets the language
cookie on language selection. The plugin should also be able to detect
language automatically (there should be no need for the code that is
currently manually setting request.lang) as it does on AMF, however
this isn't currently working for reasons unknown, so the request.lang
related logic has been left in place, for now.

The code around request.lang has been tweaked as there was a bug where
the language switch from en to cy wouldn't happen on the first try.
It might be worth considering whether the onRender hook is appropriate
for i18next.init VS running it only once with initFrontend(). The same
could be said about nunjucks configuration.