OLH-3286: Fix cookie banner
Correct cookie domain parameter for the analytics script, which also
contains cookie banner logic. Previously the banner would continue to
reappear due to the cookie_preferences_set cookie not actually being
set. Now the current cookie preference selection should be reflected in
the value of the cookie_preferences_set cookie.
Additionally, corrected the name of the "contstants" file, and added
integration test for the cookie banner.