OLH-3398: Make request for item with `outcome_id` and return `outcome`
Retrieve outcome of a journey from dynamo using the outcome_id in the
access token included with the GET request to journeyOutcome.
If the output of the get command contains an `outcome` in a valid
format, return this to the client with status code 200.
If an outcome could not be found, return a 404.
If a generic DB error occurs, return a 500.