OLH-3565: Cache getKMSKey output for JourneyOutcome
This is an optimisation for the getKmsKey util on the journey outcome
handler to avoid expensive network calls and cryptographic conversions
if they are unnecessary.
When getKmsKey is called, it now first checks that a key with the same
alias has already been stored in the kmsKeyCache. If it has not been,
it calls getPublicKey and stores the key in cache, if successful.
If it has been cached, getKmsKey should return the cached version.